# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Produtos",
    "description": "API de Produtos usando Express e MongoDB",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/produtos": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Criação de um novo Produto",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nome",
            "in": "formData",
            "description": "Nome do produto",
            "required": true,
            "type": "string"
          },
          {
            "name": "descricao",
            "in": "formData",
            "description": "Descrição do produto",
            "required": true,
            "type": "string"
          },
          {
            "name": "quantidade",
            "in": "formData",
            "description": "Quantidade do produto",
            "required": true,
            "type": "number"
          },
          {
            "name": "preco",
            "in": "formData",
            "description": "Preço do produto",
            "required": true,
            "type": "number"
          },
          {
            "name": "desconto",
            "in": "formData",
            "description": "Desconto do produto",
            "required": true,
            "type": "number"
          },
          {
            "name": "dataDesconto",
            "in": "formData",
            "description": "Data do desconto",
            "required": true,
            "type": "string"
          },
          {
            "name": "categoria",
            "in": "formData",
            "description": "Categoria do produto",
            "required": true,
            "type": "string"
          },
          {
            "name": "imagem",
            "in": "formData",
            "description": "Imagem do produto",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto criado com sucesso"
          },
          "500": {
            "description": "Um erro aconteceu"
          }
        }
      },

"get": {
        "summary": "Listagem de produtos",
        "description": "Retorna uma lista de todos os produtos cadastrados no sistema",
        "tags": ["Produtos"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          }
        }
      }
    },
    "/produtos/{id}": {
      "get": {
        "summary": "Busca produtos por ID",
        "description": "ID do produto a ser buscado",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar Produto",
        "description": "Atualiza o produto no banco de dados",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Produto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          },
          "500": {
            "description": "Um erro ocorreu"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Remoção de um Produto",
        "description": "Exclui um produto do banco de dados utilizando o ID",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          },
          "500": {
            "description": "Um erro ocorreu"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      }
    },
    "/buscanome": {
      "get": {
        "summary": "Busca de nome",
        "description": "Busca o nome do produto (case insensitive)",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto achado com sucesso"
          },
          "404": {
            "description": "Produto com este nome não encontrado"
          },
          "500": {
            "description": "Erro desconhecido"
          }
        }
      }
    },
    "/buscacategoria": {
      "get": {
        "summary": "Busca por Categoria",
        "description": "Busca um produto que se enquadra numa categoria específica. (case insensitive)",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "categoria",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto achado com sucesso"
          },
          "404": {
            "description": "Produto não encontrado com esta categoria"
          },
          "500": {
            "description": "Erro desconhecido"
          }
        }
      }
    },
    "/buscaId": {
      "get": {
        "summary": "Busca pelo Id e nome",
        "description": "Busca o produto pelo seu id e nome",
        "tags": ["Produtos"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto achado com sucesso"
          },
          "404": {
            "description": "ID inválido"
          },
          "500": {
            "description": "O parâmetro 'nome' deve ser uma string"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Produto": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "quantidade": { "type": "integer" },
          "preco": { "type": "integer" },
          "desconto": { "type": "integer" },
          "dataDesconto": { "type": "string" },
          "categoria": { "type": "string" },
          "imagem": {
            "type": "string",
            "description": "A imagem será uma imagem real que será convertida em uma string no banco de dados"
          }
        }
      }
    }
  }
}